refactor: Divide Frame
인터페이스와 구현이 같이 있는 추상 클래스였던 Frame을 인터페이스만
제공하는 Frame과 중복 코드에 대한 구현을 제공하는 DefaultFrame으로 나눔
